mails: Move Email SQL model to new `pcapi.core.mails` sub-package
